refactor of store
